requirements.txt
Flask==2.2.3
Markupsafe==2.1.2
Werkzeug==2.2.3
itsdangerous==2.1.2
psutil==5.8.0
plotly==5.3.0
tenacity==8.0.1
boto==2.49.0
kubernetes==10.0.1